Removed most localization files
