--- Source/Basic-Conditions-And-Behaviors/Editor/UI/Drawers/Audio Data Drawers/AudioDataLocalizedStringDrawer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 91936ac9d9704e928c2160f6a91e4f7b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
