--- Source/Basic-Conditions-And-Behaviors/Editor/UI/Drawers/Audio Data Drawers/ResourceAudioDataLocalizedStringDrawer.cs
@@ -1,11 +1,0 @@
-﻿namespace VRBuilder.Editor.UI.Drawers
-{
-    /// <summary>
-    /// Custom drawer for localized strings in `Play Audio File` behavior's audio data to flatten visible hierarchy.
-    /// </summary>
-    public class ResourceAudioDataLocalizedStringDrawer : AudioDataLocalizedStringDrawer
-    {
-        /// <inheritdoc />
-        protected override string defaultValueName { get { return "Default resource path"; } }
-    }
-}
