--- Source/Basic-Conditions-And-Behaviors/Editor/UI/Drawers/Audio Data Drawers/ResourceAudioDataLocalizedStringDrawer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2dac7e6a8cc54cf49933cb5846774842
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
