--- Source/Basic-Conditions-And-Behaviors/Editor/UI/Drawers/Audio Data Drawers/TextToSpeechAudioDataLocalizedStringDrawer.cs
@@ -1,11 +1,0 @@
-﻿namespace VRBuilder.Editor.UI.Drawers
-{
-    /// <summary>
-    /// Custom drawer for localized strings in `Play TTS Audio` behavior's audio data to flatten visible hierarchy.
-    /// </summary>
-    public class TextToSpeechAudioDataLocalizedStringDrawer : AudioDataLocalizedStringDrawer
-    {
-        /// <inheritdoc />
-        protected override string defaultValueName { get { return "Default text"; } }
-    }
-}
