--- Source/Basic-UI-Component/Runtime/CourseController/Controllers/ILocalizationProvider.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6c87bd5d92274892ba02e7a4f80bba03
-timeCreated: 1614678628
