+++ Source/Core/Runtime/Localization/LanguageUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e94a5fb21b430a6428d2052bdc54cf97
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
