--- Source/Core/Runtime/Localization/Localization.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 2d074cad66073fc4ca4184bd447fa9b7
-timeCreated: 1502090729
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
