--- Source/Core/Runtime/Localization/LocalizationConfig.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 52e9745f7453463fa4ba059abd2fa206
-timeCreated: 1614676688
