--- Source/Core/Runtime/Localization/LocalizationReader.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ebc5e77ad0df4b8d9d96f7b0e4ae7672
-timeCreated: 1614762924
