--- Source/Core/Runtime/Localization/LocalizationUtils.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4d68a7c921dd49bf9939d42ac4f656c2
-timeCreated: 1544104285
