--- Source/Core/Runtime/Localization/LocalizedString.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d7548200c1da3ef4f89f3e3085359086
-timeCreated: 1502109841
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
