--- Source/Core/Runtime/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83e859395be5a4c4c9557b7acc7bbea0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
