--- Source/Core/Runtime/Serialization/NewtonsoftJsonSerializer/LocalizedStringConverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a61763a491721cb4d8f14c84110dd786
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
