--- Source/Core/Tests/Utils/Mocks/Behaviors/LocalizedStringBehaviorMock.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef0f7cbbe5944534d9bb5a04a0f73f58
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
